requirements.txt
torch
scikit-learn
numpy
pandas
diffusers
datasets
transformers
jupyter
huggingface_hub
accelerate
einops
huggingface-hub
jupyter
matplotlib
numpy
opencv-python
peft
pillow
scipy
torchvision
tqdm
uuid
torch_pca
